update is flink job check

+++ service/flink/jar.go
@@ -3,6 +3,7 @@ package flink
 import (
 	"context"
 	"fmt"
+	"github.com/DataWorkbench/common/qerror"
 	"github.com/DataWorkbench/gproto/pkg/model"
 	"strings"
 
@@ -26,13 +27,27 @@ func NewJarExecutor(bm *BaseExecutor, ctx context.Context) *JarExecutor {
 func (jarExec *JarExecutor) Run(ctx context.Context, info *request.RunJob) (*zeppelin.ParagraphResult, error) {
 	var result *zeppelin.ParagraphResult
 	var noteId string
+	var err error
 	defer func() {
-		if result != nil && result.Status != zeppelin.RUNNING && len(noteId) > 0 {
-			_ = jarExec.zeppelinClient.DeleteNote(noteId)
+		if err == nil {
+			if result != nil && result.Status != zeppelin.RUNNING && len(noteId) > 0 {
+				if result.Status == zeppelin.ERROR && len(result.Results) > 0 {
+					for _, re := range result.Results {
+						if strings.EqualFold(re.Type, "TEXT") && strings.Contains(re.Data, "Caused by: java.net.ConnectException: Connection refused") {
+							err = qerror.FlinkRestError
+						}
+					}
+				}
+				_ = jarExec.zeppelinClient.DeleteNote(noteId)
+			} else if (result.Status == zeppelin.RUNNING || result.Status == zeppelin.FINISHED) &&
+				len(result.JobId) != 32 && len(noteId) > 0 {
+				result.Status = zeppelin.ABORT
+				_ = jarExec.zeppelinClient.DeleteNote(noteId)
+			}
 		}
 	}()
 
-	result, err := jarExec.preCheck(ctx, info.InstanceId)
+	result, err = jarExec.preCheck(ctx, info.InstanceId)
 	if err != nil {
 		return nil, err
 	} else if result != nil {
@@ -121,8 +136,8 @@ func (jarExec *JarExecutor) Cancel(ctx context.Context, instanceId string, space
 	return jarExec.cancelJob(ctx, instanceId, spaceId, clusterId)
 }
 
-func (jarExec *JarExecutor) Release(ctx context.Context,instanceId string) error{
-	return jarExec.release(ctx,instanceId)
+func (jarExec *JarExecutor) Release(ctx context.Context, instanceId string) error {
+	return jarExec.release(ctx, instanceId)
 }
 
 func (jarExec *JarExecutor) Validate(jobCode *model.StreamJobCode) (bool, string, error) {
